Fix time index so that speed and LFPs have same time
For certain datasets, there are multiple DSPs which results in different timing.  `get_interpolated_position_dataframe` automatically interpolates to the time of all the LFPs concatenated. If only loading a subset of tetrodes, these need to be reindexed to this time so that they match. Alternatively, we could provide a different time function to `get_interpolated_position_dataframe`.